[localevaluation] try to be on the safe side when computing orders

+++ dune/gdt/playground/localevaluation/elliptic.hh
@@ -62,8 +62,9 @@ public:
                const Stuff::LocalfunctionSetInterface<E, D, d, R, rT, rCT>& testBase,
                const Stuff::LocalfunctionSetInterface<E, D, d, R, rA, rCA>& ansatzBase) const
   {
-    return local_diffusion_factor.order() + local_diffusion_tensor.order() + std::max(int(testBase.order() - 1), 0)
-           + std::max(int(ansatzBase.order() - 1), 0);
+    return local_diffusion_factor.order() + local_diffusion_tensor.order()
+           + std::max(ssize_t(testBase.order()) - 1, ssize_t(0))
+           + std::max(ssize_t(ansatzBase.order()) - 1, ssize_t(0));
   } // ... order(...)
 
   /**
